+++ functions/src/index.ts
@@ -3,7 +3,7 @@ import * as express from 'express'
 import * as cors from 'cors'
 import { OdaiCurrentParams, OdaiPostRequestParams } from './odai/Odai'
 import { OdaiRepository, OdaiRepositoryImpl } from './odai/OdaiRepository'
-import { OdaiUseCase, OdaiUseCaseImpl } from './odai/OdaiUseCase'
+import { OdaiService, OdaiServiceImpl } from './odai/OdaiService'
 
 const REGION = 'asia-northeast1'
 
@@ -11,7 +11,7 @@ const app = express()
 app.use(cors({ origin: true }))
 
 const odaiRepository: OdaiRepository = new OdaiRepositoryImpl()
-const odaiUseCase: OdaiUseCase = new OdaiUseCaseImpl(odaiRepository)
+const odaiService: OdaiService = new OdaiServiceImpl(odaiRepository)
 
 const errorResponse = (res: express.Response, statusCode: number, message: string) => {
   return res.status(statusCode).send({ error: true, message })
@@ -22,7 +22,7 @@ app.post('/odai', async (req: express.Request, res) => {
   if (!slackTeamId || !title || !createdBy) {
     return errorResponse(res, 422, 'Illegal Argument')
   }
-  const result = await odaiUseCase.create({ slackTeamId, title, createdBy })
+  const result = await odaiService.create({ slackTeamId, title, createdBy })
   if (result === 'error') {
     return errorResponse(res, 500, 'Internal Server Error')
   }
@@ -37,7 +37,7 @@ app.get('/odai/current', async (req: express.Request, res) => {
   if (!slackTeamId) {
     return errorResponse(res, 422, 'Illegal Argument')
   }
-  const result = await odaiUseCase.getCurrent({ slackTeamId })
+  const result = await odaiService.getCurrent({ slackTeamId })
   return res.send({ odai: result })
 })
