+++ src/index.ts
@@ -1,7 +1,7 @@
 import { App } from '@slack/bolt'
 import * as dotenv from 'dotenv'
-import { createKotae } from './actions/kotae'
-import { createOdai } from './actions/odai'
+import { createKotae } from './kotae/KotaeAction'
+import { createOdai } from './odai/OdaiAction'
 
 dotenv.config()
