+++ src/odai/OdaiUseCase.ts
@@ -1,7 +1,7 @@
-import { api } from '../axios'
-import { OdaiPostRequestParams, OdaiPostResponse } from '../types/Odai'
+import { api } from '../api/axios'
+import { OdaiPostRequestParams, OdaiPostResponse } from './Odai'
 
-export class OdaiService {
+export class OdaiUseCase {
   async create(data: OdaiPostRequestParams) {
     const result = await api.post<OdaiPostResponse>('/odai', data)
     return result.data
